add session in posts file

+++ posts.mjs
@@ -1,40 +1,72 @@
 import { Router } from "express";
-import { postsCollection } from "./database.mjs";
+import { client, postsCollection, usersCollection } from "./database.mjs";
 import { ObjectId } from "mongodb";
 const router = Router();
 
-router.get("/", async (req, res, next) => {
+router.get("/:_id", async (req, res, next) => {
   try {
-    const posts = await postsCollection.find({}).toArray();
-    console.log(posts);
-    res.status(200).send(posts);
+    const post = await postsCollection.findOne({
+      _id: new ObjectId(req.params._id),
+    });
+    res.status(200).send(post);
   } catch (error) {
     next(error);
   }
 });
 
 router.post("/", async (req, res, next) => {
+  const session = client.startSession();
   try {
+    session.startTransaction();
     const data = req.body;
-    const insertedPost = await postsCollection.insertOne(data);
-    // update user posts list
+    const insertedPost = await postsCollection.insertOne(data, { session });
+    await usersCollection.updateOne(
+      { _id: data.owner },
+      { $push: { posts: insertedPost.insertedId } },
+      { session }
+    );
     data["_id"] = insertedPost.insertedId;
+    await session.commitTransaction();
     res.status(201).send(data);
   } catch (error) {
+    await session.abortTransaction();
     next(error);
+  } finally {
+    await session.endSession();
   }
 });
 
 router.delete("/:_id", async (req, res, next) => {
+  const session = client.startSession();
+
   try {
-    const deleted = await postsCollection.deleteOne({
+    session.startTransaction();
+
+    const postToDelete = await postsCollection.findOne({
       _id: new ObjectId(req.params._id),
     });
+
+    if (!postToDelete) throw new Error("Could not find post.");
+
+    const deleted = await postsCollection.deleteOne(
+      {
+        _id: new ObjectId(req.params._id),
+      },
+      { session }
+    );
     if (!deleted.deletedCount) throw new Error("Could not delete.");
-    // remove from posts list in the user document
+    await usersCollection.updateOne(
+      { _id: postToDelete.owner },
+      { $pop: { posts: req.params._id } },
+      { session }
+    );
+    await session.commitTransaction();
     res.status(200).send("Deleted");
   } catch (error) {
+    await session.abortTransaction();
     next(error);
+  } finally {
+    await session.endSession();
   }
 });
